+++ packages/client/src/index.ts
@@ -1,5 +1,6 @@
 export * from "./client"
 export * from "./re-exports"
+export * from "./descriptors"
 export type { EventPhase } from "./event"
 export type { PolkadotSigner } from "@polkadot-api/polkadot-signer"
 export type { PolkadotClient, TypedApi, FixedSizeArray } from "./types"