+++ packages/utils/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- feat: add `filterObject` utility
+
 ## 0.0.1 - 2024-04-03
 
 ### Changed